+++ core/src/main/java/net/infumia/frame/view/ViewCreatorImpl.java
@@ -1,17 +1,14 @@
 package net.infumia.frame.view;
 
-import java.lang.reflect.Constructor;
 import java.util.concurrent.CompletableFuture;
 import net.infumia.frame.injection.AnnotationAccessor;
 import net.infumia.frame.injection.InjectionRequester;
 import net.infumia.frame.injection.InjectionService;
 import net.infumia.frame.injection.InjectionServicePipeline;
-import net.infumia.frame.injector.InjectionRequest;
-import net.infumia.frame.injector.Injector;
+import net.infumia.frame.injector.InjectorNoArg;
 import net.infumia.frame.injector.InjectorRegistry;
 import net.infumia.frame.util.Preconditions;
 import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
 
 public final class ViewCreatorImpl implements ViewCreator {
 
@@ -20,7 +17,7 @@ public final class ViewCreatorImpl implements ViewCreator {
     private final InjectionRequester<Object> requester;
 
     public ViewCreatorImpl() {
-        this.injectors = InjectorRegistry.create().register(new NoArg());
+        this.injectors = InjectorRegistry.create().register(new InjectorNoArg());
         this.pipeline = InjectionServicePipeline.create(InjectionService.create(this.injectors));
         this.requester = InjectionRequester.create(this.pipeline);
     }
@@ -57,26 +54,5 @@ public final class ViewCreatorImpl implements ViewCreator {
                     viewClass
                 );
             });
-    }
-
-    private static final class NoArg implements Injector<Object> {
-
-        @Nullable
-        @Override
-        public Object inject(@NotNull final InjectionRequest<Object> ctx) {
-            final Class<?> viewClass = ctx.injectedClass();
-            Constructor<?> constructor = null;
-            boolean old = false;
-            try {
-                constructor = viewClass.getConstructor();
-                old = constructor.isAccessible();
-                return constructor.newInstance();
-            } catch (final Exception ignored) {} finally {
-                if (constructor != null) {
-                    constructor.setAccessible(old);
-                }
-            }
-            return null;
-        }
     }
 }
